feat: Header New component -> all screens
