+++ components/Header/index.js
@@ -1,0 +1,3 @@
+import Header from './Header';
+
+export default Header;
